requirements.txt
onnx>=1.20.0
google-generativeai>=0.8.0
numpy>=1.23.2
onnxruntime>=1.16.0  # Optional: for numerical output comparison in model_comparator.py

